Move helpers out of graphic and into util package
